+++ src/ui/components/graph_panel.h
@@ -9,7 +9,7 @@ struct GraphPanel {
     Vec2 scroll_pos;
     f32 node_properties_panel_height;
     f32 width;
-    int cur_node; // Index of current node
+    size_t cur_node; // Index of current node
     PapayaSlot* dragged_slot;
     PapayaSlot* displaced_slot;
 };
